// src/types/dom.ts
// Типы для DOM элементов страницы
export type RegistrationPageElements = {
    form: HTMLFormElement;
    lastNameInput: HTMLInputElement;
    firstNameInput: HTMLInputElement;
    fatherNameInput: HTMLInputElement;
    emailInput: HTMLInputElement;
    passwordInput: HTMLInputElement;
    repeatPasswordInput: HTMLInputElement;
    submitButton: HTMLButtonElement;
    errorContainer: HTMLElement | null;
  };